Spinner: reduced gap between up and down arrows, which was increased by previous commit (issue #329)

+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatUIUtils.java
@@ -660,7 +660,7 @@ public class FlatUIUtils
 	 * @since 1.1
 	 */
 	public static void paintArrow( Graphics2D g, int x, int y, int width, int height,
-		int direction, boolean chevron, int arrowSize, int xOffset, int yOffset )
+		int direction, boolean chevron, int arrowSize, float xOffset, float yOffset )
 	{
 		// compute arrow width/height
 		int aw = UIScale.scale( arrowSize + (chevron ? 0 : 1) );
@@ -679,8 +679,8 @@ public class FlatUIUtils
 		int extra = chevron ? 1 : 0;
 
 		// compute arrow location
-		float ox = ((width - (aw + extra)) / 2f) + UIScale.scale( (float) xOffset );
-		float oy = ((height - (ah + extra)) / 2f) + UIScale.scale( (float) yOffset );
+		float ox = ((width - (aw + extra)) / 2f) + UIScale.scale( xOffset );
+		float oy = ((height - (ah + extra)) / 2f) + UIScale.scale( yOffset );
 		int ax = x + ((direction == SwingConstants.WEST) ? -Math.round( -ox ) : Math.round( ox ));
 		int ay = y + ((direction == SwingConstants.NORTH) ? -Math.round( -oy ) : Math.round( oy ));
